# internal/models/subscription.go
package models

import "github.com/google/uuid"

type baseStruct struct {
	Name      string    `json:"name"`
	Price     int       `json:"price"`
	UserId    uuid.UUID `json:"user_id"`
	StartDate string    `json:"start_date"`
	EndDate   string    `json:"end_date"`
}

type SubscriptionRequest struct {
	baseStruct
}

type Subscription struct {
	Id uuid.UUID `json:"id"`
	baseStruct
}

type SubscriptionPriceSum struct {
	Sum int `json:"sum"`
}